+++ db/db_merge_test.cc
@@ -32,6 +32,8 @@ class DBMergeTest : public testing::Test {
  public:
   DBMergeTest() {
     options_.create_if_missing = true;
+    options_.write_buffer_manager.reset(
+        new WriteBufferManager(options_.db_write_buffer_size));
     // avoid stalling the tests.
     options_.disable_write_stall = true;
     options_.avoid_flush_during_shutdown = true;