+++ options/db_options.h
@@ -59,6 +59,7 @@ struct ImmutableDBOptions {
   bool advise_random_on_open;
   double experimental_mempurge_threshold;
   size_t db_write_buffer_size;
+  std::shared_ptr<WriteBufferManager> write_buffer_manager;
   DBOptions::AccessHint access_hint_on_compaction_start;
   bool new_table_reader_for_compaction_inputs;
   size_t random_access_max_buffer_size;